obs-ffmpeg: Split ffmpeg_muxer output file by size or time
This commit adds two new properties to split output file in the output
`ffmpeg_muxer`.
- `max_time_sec` specifies the limit in seconds.
- `max_size_mb` specifies the limit in megabytes.
If each property is non-positive, the associated limit is disabled.
When splitting the file, the same properties as replay buffer are taken
to generate the new file name.
Another output ffmpeg_mpegts_muxer shares the code but is not affected
since the output is used only for streaming.