libobs: Preserve audio mixers on unavailable source
Try to preserve audio mixer settings in a scene collection even if the
source does not have the flag `OBS_SOURCE_AUDIO`.
The audio mixer settings became 0 when saving unavailable source because
the flag is not set. This unexpectedly changes audio mixer settings when
temporarily remove an audio source plugin. Then install the plugin, the
audio does not go to the mixer.